requirements.txt
cohere==5.15.0
fastapi==0.115.12
langchain==0.3.26
mistralai==1.8.2
openai==1.82.1
pydantic==2.11.5
python-dotenv==1.1.0
qdrant-client==1.14.3
python-multipart
uvicorn[standard]
PyPDF2~=3.0.1
streamlit~=1.45.1
pymongo~=4.13.2
